Implement byte pattern search in the process memory of KO
This PR implements a utility class called MAPPED_MEMORY that allows
copying large chunks of memory from an external process.
It allows the user to perform searches for patterns of byte in the
mapped memory and returns the address directly as expressed in the
address space of the external process.

This is used to search for the pointers for skill cooldowns, player hp
and mp in the KO memory. These values are read in the construction of
KO_CLIENT and displayed dynamically in a simple node that prints them to
the terminal.

Closes:SW-29